third_party/WebKit/LayoutTests/platform/linux/virtual/mojo-localstorage/external/wpt/webstorage/storage_string_conversion-expected.txt

This is a testharness.js-based test.
FAIL localStorage only stores strings assert_equals: expected "function(){}" but got "function (){}"
FAIL sessionStorage only stores strings assert_equals: expected "function(){}" but got "function (){}"
Harness: the test ran to completion.
